Patch: Shortcut fixes on 2.0.3

+++ src/KenjinxAndroid/app/src/main/java/org/kenjinx/android/viewmodels/MainViewModel.kt
@@ -9,6 +9,7 @@ import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.sync.Semaphore
 import org.kenjinx.android.GameController
 import org.kenjinx.android.GameHost
+import org.kenjinx.android.KenjinxNative
 import org.kenjinx.android.Logging
 import org.kenjinx.android.MainActivity
 import org.kenjinx.android.MotionSensorManager
@@ -17,7 +18,6 @@ import org.kenjinx.android.NativeHelpers
 import org.kenjinx.android.PerformanceManager
 import org.kenjinx.android.PhysicalControllerManager
 import org.kenjinx.android.RegionCode
-import org.kenjinx.android.KenjinxNative
 import org.kenjinx.android.SystemLanguage
 import org.kenjinx.android.UiHandler
 import java.io.File
@@ -30,13 +30,16 @@ class MainViewModel(val activity: MainActivity) {
     var controller: GameController? = null
     var performanceManager: PerformanceManager? = null
     var selected: GameModel? = null
+
     val loadGameModel: MutableState<GameModel?> = mutableStateOf<GameModel?>(null)
     val bootPath: MutableState<String?> = mutableStateOf<String?>(null)
-    val forceNceAndPptc: MutableState<Boolean> = mutableStateOf<Boolean>(false)
+    val forceNceAndPptc: MutableState<Boolean> = mutableStateOf(false)
+
     var isMiiEditorLaunched = false
     val userViewModel = UserViewModel()
     val logging = Logging(this)
     var firmwareVersion = ""
+
     private var gameTimeState: MutableState<Double>? = null
     private var gameFpsState: MutableState<Double>? = null
     private var fifoState: MutableState<Double>? = null
@@ -53,8 +56,8 @@ class MainViewModel(val activity: MainActivity) {
             field = value
             field?.setProgressStates(showLoading, progressValue, progress)
         }
-    var navController: NavHostController? = null
 
+    var navController: NavHostController? = null
     var homeViewModel: HomeViewModel = HomeViewModel(activity, this)
 
     init {
@@ -74,19 +77,19 @@ class MainViewModel(val activity: MainActivity) {
         motionSensorManager?.setControllerId(-1)
     }
 
-    fun loadGame(game: GameModel, overrideSettings: Boolean? = false, forceNceAndPptc: Boolean? = false): Int {
+    fun loadGame(
+        game: GameModel,
+        overrideSettings: Boolean? = false,
+        forceNceAndPptc: Boolean? = false
+    ): Int {
         KenjinxNative.jnaInstance.deviceReinitEmulation()
         MainActivity.mainViewModel?.activity?.uiHandler = UiHandler()
 
         val descriptor = game.open()
-
-        if (descriptor == 0)
-            return 0
+        if (descriptor == 0) return 0
 
         val update = game.openUpdate()
-
-        if(update == -2)
-        {
+        if (update == -2) {
             return -2
         }
 
@@ -94,10 +97,8 @@ class MainViewModel(val activity: MainActivity) {
         isMiiEditorLaunched = false
 
         val settings = QuickSettings(activity)
-
-        if(overrideSettings == true)
-        {
-            settings.overrideSettings(forceNceAndPptc);
+        if (overrideSettings == true) {
+            settings.overrideSettings(forceNceAndPptc)
         }
 
         var success = KenjinxNative.jnaInstance.graphicsInitialize(
@@ -108,16 +109,14 @@ class MainViewModel(val activity: MainActivity) {
             maxAnisotropy = settings.maxAnisotropy,
             backendThreading = org.kenjinx.android.BackendThreading.Auto.ordinal
         )
-
-        if (!success)
-            return 0
+        if (!success) return 0
 
         val nativeHelpers = NativeHelpers.instance
-        val nativeInterop = NativeGraphicsInterop()
-
-        nativeInterop.VkRequiredExtensions = arrayOf("VK_KHR_surface", "VK_KHR_android_surface")
-        nativeInterop.VkCreateSurface = nativeHelpers.getCreateSurfacePtr()
-        nativeInterop.SurfaceHandle = 0
+        val nativeInterop = NativeGraphicsInterop().apply {
+            VkRequiredExtensions = arrayOf("VK_KHR_surface", "VK_KHR_android_surface")
+            VkCreateSurface = nativeHelpers.getCreateSurfacePtr()
+            SurfaceHandle = 0
+        }
 
         val driverViewModel = VulkanDriverViewModel(activity)
         val drivers = driverViewModel.getAvailableDrivers()
@@ -125,22 +124,18 @@ class MainViewModel(val activity: MainActivity) {
 
         if (driverViewModel.selected.isNotEmpty()) {
             val metaData = drivers.find { it.driverPath == driverViewModel.selected }
-
             metaData?.apply {
                 val privatePath = activity.filesDir
                 val privateDriverPath = privatePath.canonicalPath + "/driver/"
                 val pD = File(privateDriverPath)
-                if (pD.exists())
-                    pD.deleteRecursively()
-
+                if (pD.exists()) pD.deleteRecursively()
                 pD.mkdirs()
 
                 val driver = File(driverViewModel.selected)
                 val parent = driver.parentFile
                 if (parent != null) {
                     for (file in parent.walkTopDown()) {
-                        if (file.absolutePath == parent.absolutePath)
-                            continue
+                        if (file.absolutePath == parent.absolutePath) continue
                         file.copyTo(File(privateDriverPath + file.name), true)
                     }
                 }
@@ -151,24 +146,21 @@ class MainViewModel(val activity: MainActivity) {
                     this.libraryName
                 )
             }
-
         }
 
         val extensions = nativeInterop.VkRequiredExtensions
-
         success = KenjinxNative.jnaInstance.graphicsInitializeRenderer(
             extensions!!,
             extensions.size,
             driverHandle
         )
-        if (!success)
-            return 0
+        if (!success) return 0
 
         val semaphore = Semaphore(1, 0)
         runBlocking {
             semaphore.acquire()
             launchOnUiThread {
-                // We are only able to initialize the emulation context on the main thread
+                // Emulation-Kontext muss im Main-Thread initialisiert werden
                 success = KenjinxNative.jnaInstance.deviceInitialize(
                     settings.memoryManagerMode.ordinal,
                     settings.useNce,
@@ -178,24 +170,20 @@ class MainViewModel(val activity: MainActivity) {
                     settings.vSyncMode.ordinal,
                     settings.enableDocked,
                     settings.enablePptc,
-                    false,
+                    /* enableLowPowerPptc */ false,
                     settings.enableFsIntegrityChecks,
                     settings.fsGlobalAccessLogMode,
                     "UTC",
                     settings.ignoreMissingServices
                 )
-
                 semaphore.release()
             }
             semaphore.acquire()
             semaphore.release()
         }
-
-        if (!success)
-            return 0
+        if (!success) return 0
 
         success = KenjinxNative.jnaInstance.deviceLoadDescriptor(descriptor, game.type.ordinal, update)
-
         return if (success) 1 else 0
     }
 
@@ -204,7 +192,6 @@ class MainViewModel(val activity: MainActivity) {
         isMiiEditorLaunched = true
 
         val settings = QuickSettings(activity)
-
         var success = KenjinxNative.jnaInstance.graphicsInitialize(
             enableMacroHLE = settings.enableMacroHLE,
             enableShaderCache = settings.enableShaderCache,
@@ -213,40 +200,33 @@ class MainViewModel(val activity: MainActivity) {
             maxAnisotropy = settings.maxAnisotropy,
             backendThreading = org.kenjinx.android.BackendThreading.Auto.ordinal
         )
-
-        if (!success)
-            return false
+        if (!success) return false
 
         val nativeHelpers = NativeHelpers.instance
-        val nativeInterop = NativeGraphicsInterop()
-
-        nativeInterop.VkRequiredExtensions = arrayOf("VK_KHR_surface", "VK_KHR_android_surface")
-        nativeInterop.VkCreateSurface = nativeHelpers.getCreateSurfacePtr()
-        nativeInterop.SurfaceHandle = 0
+        val nativeInterop = NativeGraphicsInterop().apply {
+            VkRequiredExtensions = arrayOf("VK_KHR_surface", "VK_KHR_android_surface")
+            VkCreateSurface = nativeHelpers.getCreateSurfacePtr()
+            SurfaceHandle = 0
+        }
 
         val driverViewModel = VulkanDriverViewModel(activity)
         val drivers = driverViewModel.getAvailableDrivers()
-
         var driverHandle = 0L
 
         if (driverViewModel.selected.isNotEmpty()) {
             val metaData = drivers.find { it.driverPath == driverViewModel.selected }
-
             metaData?.apply {
                 val privatePath = activity.filesDir
                 val privateDriverPath = privatePath.canonicalPath + "/driver/"
                 val pD = File(privateDriverPath)
-                if (pD.exists())
-                    pD.deleteRecursively()
-
+                if (pD.exists()) pD.deleteRecursively()
                 pD.mkdirs()
 
                 val driver = File(driverViewModel.selected)
                 val parent = driver.parentFile
                 if (parent != null) {
                     for (file in parent.walkTopDown()) {
-                        if (file.absolutePath == parent.absolutePath)
-                            continue
+                        if (file.absolutePath == parent.absolutePath) continue
                         file.copyTo(File(privateDriverPath + file.name), true)
                     }
                 }
@@ -257,24 +237,20 @@ class MainViewModel(val activity: MainActivity) {
                     this.libraryName
                 )
             }
-
         }
 
         val extensions = nativeInterop.VkRequiredExtensions
-
         success = KenjinxNative.jnaInstance.graphicsInitializeRenderer(
             extensions!!,
             extensions.size,
             driverHandle
         )
-        if (!success)
-            return false
+        if (!success) return false
 
         val semaphore = Semaphore(1, 0)
         runBlocking {
             semaphore.acquire()
             launchOnUiThread {
-                // We are only able to initialize the emulation context on the main thread
                 success = KenjinxNative.jnaInstance.deviceInitialize(
                     settings.memoryManagerMode.ordinal,
                     settings.useNce,
@@ -284,45 +260,36 @@ class MainViewModel(val activity: MainActivity) {
                     settings.vSyncMode.ordinal,
                     settings.enableDocked,
                     settings.enablePptc,
-                    false,
+                    /* enableLowPowerPptc */ false,
                     settings.enableFsIntegrityChecks,
                     settings.fsGlobalAccessLogMode,
                     "UTC",
                     settings.ignoreMissingServices
                 )
-
                 semaphore.release()
             }
             semaphore.acquire()
             semaphore.release()
         }
+        if (!success) return false
 
-        if (!success)
-            return false
-
-        success = KenjinxNative.jnaInstance.deviceLaunchMiiEditor()
-
-        return success
+        return KenjinxNative.jnaInstance.deviceLaunchMiiEditor()
     }
 
     fun clearPptcCache(titleId: String) {
         if (titleId.isNotEmpty()) {
             val basePath = MainActivity.AppPath + "/games/$titleId/cache/cpu"
             if (File(basePath).exists()) {
-                var caches = mutableListOf<String>()
-
+                val caches = mutableListOf<String>()
                 val mainCache = basePath + "${File.separator}0"
                 File(mainCache).listFiles()?.forEach {
-                    if (it.isFile && it.name.endsWith(".cache"))
-                        caches.add(it.absolutePath)
+                    if (it.isFile && it.name.endsWith(".cache")) caches.add(it.absolutePath)
                 }
                 val backupCache = basePath + "${File.separator}1"
                 File(backupCache).listFiles()?.forEach {
-                    if (it.isFile && it.name.endsWith(".cache"))
-                        caches.add(it.absolutePath)
+                    if (it.isFile && it.name.endsWith(".cache")) caches.add(it.absolutePath)
                 }
-                for (path in caches)
-                    File(path).delete()
+                for (path in caches) File(path).delete()
             }
         }
     }
@@ -331,17 +298,14 @@ class MainViewModel(val activity: MainActivity) {
         if (titleId.isNotEmpty()) {
             val basePath = MainActivity.AppPath + "/games/$titleId/cache/shader"
             if (File(basePath).exists()) {
-                var caches = mutableListOf<String>()
+                val caches = mutableListOf<String>()
                 File(basePath).listFiles()?.forEach {
-                    if (!it.isFile)
-                        it.delete()
-                    else {
-                        if (it.name.endsWith(".toc") || it.name.endsWith(".data"))
-                            caches.add(it.absolutePath)
+                    if (!it.isFile) it.delete()
+                    else if (it.name.endsWith(".toc") || it.name.endsWith(".data")) {
+                        caches.add(it.absolutePath)
                     }
                 }
-                for (path in caches)
-                    File(path).delete()
+                for (path in caches) File(path).delete()
             }
         }
     }
@@ -388,21 +352,13 @@ class MainViewModel(val activity: MainActivity) {
         gameFps: Double,
         gameTime: Double
     ) {
-        fifoState?.apply {
-            this.value = fifo
-        }
-        gameFpsState?.apply {
-            this.value = gameFps
-        }
-        gameTimeState?.apply {
-            this.value = gameTime
-        }
+        fifoState?.let { it.value = fifo }
+        gameFpsState?.let { it.value = gameFps }
+        gameTimeState?.let { it.value = gameTime }
+
         usedMemState?.let { usedMem ->
             totalMemState?.let { totalMem ->
-                MainActivity.performanceMonitor.getMemoryUsage(
-                    usedMem,
-                    totalMem
-                )
+                MainActivity.performanceMonitor.getMemoryUsage(usedMem, totalMem)
             }
         }
         frequenciesState?.let { MainActivity.performanceMonitor.getFrequencies(it) }
@@ -415,8 +371,9 @@ class MainViewModel(val activity: MainActivity) {
     fun navigateToGame() {
         navController?.navigate("game")
         activity.isGameRunning = true
-        if (QuickSettings(activity).enableMotion)
+        if (QuickSettings(activity).enableMotion) {
             motionSensorManager?.register()
+        }
     }
 
     fun setProgressStates(
